requirements.txt
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.0.0
pandas>=2.0.0
matplotlib>=3.0.0
numpy>=1.20.0